requirements.txt
pyktok
playwright
opencv-python
easyocr
numpy
pillow
requests
python-dotenv
google-cloud-vision
openai-whisper
googlemaps
pandas
google-generativeai